requirements.txt
blinker==1.9.0 ; python_version >= "3.13"
cachetools==5.5.2 ; python_version >= "3.13"
certifi==2025.1.31 ; python_version >= "3.13"
charset-normalizer==3.4.1 ; python_version >= "3.13"
click==8.1.8 ; python_version >= "3.13"
colorama==0.4.6 ; python_version >= "3.13" and platform_system == "Windows"
flask==3.1.0 ; python_version >= "3.13"
google-api-core==2.24.2 ; python_version >= "3.13"
google-api-python-client==2.166.0 ; python_version >= "3.13"
google-auth-httplib2==0.2.0 ; python_version >= "3.13"
google-auth-oauthlib==1.2.1 ; python_version >= "3.13"
google-auth==2.38.0 ; python_version >= "3.13"
googleapis-common-protos==1.69.2 ; python_version >= "3.13"
gspread==6.2.0 ; python_version >= "3.13"
gunicorn==23.0.0 ; python_version >= "3.13"
httplib2==0.22.0 ; python_version >= "3.13"
idna==3.10 ; python_version >= "3.13"
itsdangerous==2.2.0 ; python_version >= "3.13"
jinja2==3.1.6 ; python_version >= "3.13"
markupsafe==3.0.2 ; python_version >= "3.13"
oauth2client==4.1.3 ; python_version >= "3.13"
oauthlib==3.2.2 ; python_version >= "3.13"
packaging==24.2 ; python_version >= "3.13"
proto-plus==1.26.1 ; python_version >= "3.13"
protobuf==6.30.2 ; python_version >= "3.13"
pyasn1-modules==0.4.2 ; python_version >= "3.13"
pyasn1==0.6.1 ; python_version >= "3.13"
pyparsing==3.2.3 ; python_version >= "3.13"
python-dotenv==1.1.0 ; python_version >= "3.13"
requests-oauthlib==2.0.0 ; python_version >= "3.13"
requests==2.32.3 ; python_version >= "3.13"
rsa==4.2 ; python_version >= "3.13"
six==1.17.0 ; python_version >= "3.13"
uritemplate==4.1.1 ; python_version >= "3.13"
urllib3==2.3.0 ; python_version >= "3.13"
werkzeug==3.1.3 ; python_version >= "3.13"
